test: Refactors tests to break up Release tests
