+++ tests/unit_tests/release/__init__.py
@@ -1,0 +1,1 @@
+"""Test the code for the Release class."""
